Refactor: extract FakeConfiguration class
