+++ src/test/kotlin/com/gmaur/investment/robotadvisor/FakeConfiguration.kt
@@ -1,0 +1,16 @@
+package com.gmaur.investment.robotadvisor
+
+import com.gmaur.investment.robotadvisor.domain.PortfolioRebalancer
+import org.mockito.Mockito
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class FakeConfiguration {
+    private val portfolioRebalancer = Mockito.mock(PortfolioRebalancer::class.java)
+
+    @Bean
+    fun rebalancer(): PortfolioRebalancer {
+        return portfolioRebalancer
+    }
+}
